fix : clearing post-user relation

+++ BE/controllers/feed.js
@@ -132,11 +132,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // if (post.creator.name !== 'Maximilian') {
-      //   const error = new Error('Not authorized!');
-      //   error.statusCode = 403;
-      //   throw error;
-      // }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -168,16 +168,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // if (post.creator.name !== 'Maximilian') {
-      //   const error = new Error('Not authorized!');
-      //   error.statusCode = 403;
-      //   throw error;
-      // }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      res.status(200).json({ message: 'Post deleted.' });
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Post deleted successfully!' });
     })
     .catch(err => {
       if (!err.statusCode) {
